pyproject.toml
[tool.poetry]
name = "catdel"
version = "0.1.0"
description = "A web-based catchment delineation tool."
authors = ["Ashkan Shokri <ashkan.shokri@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.10.0"
streamlit = "^1.36.0"
pysheds = "^0.4"
fiona = "^1.9.6"
matplotlib = "^3.9.0"
folium = "^0.17.0"
numpy = "1.25.2"
geopandas = "^1.0.0"
streamlit-folium = "^0.21.0"
streamlit-js-eval = "^0.1.7"
pyyaml = "^6.0.1"
streamlit-javascript = "^0.1.5"
htbuilder = "^0.6.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
replit-object-storage = "^1.0.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
